% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getComtradeDict.R
\name{getComtradeDict}
\alias{getComtradeDict}
\title{Get Comtrade dictionary}
\usage{
getComtradeDict(parameter = "region")
}
\arguments{
\item{parameter}{which parameter do you want the dictionary for? Specify "HS", "H1" etc for the product classifications (see link below)}
}
\value{
the dictionary as a data.frame
}
\description{
Get a dictionary for a parameter
}
\details{
This function downloads the data dictionary from the UN Comtrade API. See link below.
}

