\name{RComtrade-package}
\alias{RComtrade-package}
\alias{RComtrade}
\docType{package}
\title{
\packageTitle{RComtrade}
}
\description{
\packageDescription{RComtrade}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RComtrade}
\packageIndices{RComtrade}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RComtrade}

Maintainer: \packageMaintainer{RComtrade}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
